requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --annotation-style=line --output-file=requirements.txt requirements-in.txt
#


adbutils==0.11.0          # via -r requirements-in.txt, uiautomator2
annotated-types==0.7.0    # via pydantic
anyio==3.7.1              # via fastapi, starlette
anytree==2.8.0            # via -r requirements-in.txt
apkutils2==1.0.0          # via adbutils
cached-property==1.5.2    # via uiautomator2
certifi==2023.11.17       # via requests
cffi==1.16.0              # via cryptography, gevent
charset-normalizer==3.3.2  # via requests
cigam==0.0.3              # via apkutils2
click==8.1.7              # via uvicorn
cn2an==0.5.23             # via -r requirements-in.txt
colorama==0.4.6           # via click, logzero, tqdm
coloredlogs==15.0.1       # via onnxruntime
cryptography==42.0.8      # via -r requirements-in.txt
decorator==5.1.1          # via retry
deprecated==1.2.14        # via uiautomator2
deprecation==2.1.0        # via adbutils
exceptiongroup==1.2.0     # via anyio
fastapi==0.104.1          # via -r requirements-in.txt
filelock==3.13.1          # via uiautomator2
flatbuffers==23.5.26      # via onnxruntime
future==0.18.3            # via zerorpc
gevent==23.9.1            # via zerorpc
greenlet==3.0.3           # via gevent
h11==0.14.0               # via uvicorn
humanfriendly==10.0       # via coloredlogs
idna==3.6                 # via anyio, requests
inflection==0.5.1         # via -r requirements-in.txt
logzero==1.7.0            # via uiautomator2
lxml==5.0.0               # via uiautomator2
markdown-it-py==2.2.0     # via rich
mdurl==0.1.2              # via markdown-it-py
mpmath==1.3.0             # via sympy
msgpack==1.0.7            # via zerorpc
numpy==1.24.3             # via -r requirements-in.txt, onnxruntime, opencv-python, ppocr-onnx, shapely
oashya==0.0.7             # via -r requirements-in.txt
onepush==1.3.0            # via -r requirements-in.txt
onnxruntime==1.16.3       # via ppocr-onnx
opencv-python==4.7.0.72   # via -r requirements-in.txt, ppocr-onnx
packaging==20.9           # via deprecation, onnxruntime, uiautomator2
paho-mqtt==1.6.1          # via -r requirements-in.txt
pillow==10.2.0            # via ppocr-onnx, uiautomator2
ppocr-onnx==0.0.3.9       # via -r requirements-in.txt
proces==0.1.7             # via cn2an
progress==1.6             # via uiautomator2
protobuf==4.25.1          # via onnxruntime
psutil==6.1.1             # via -r requirements-in.txt
py==1.11.0                # via retry
pyclipper==1.3.0.post5    # via ppocr-onnx
pycparser==2.21           # via cffi
pycryptodome==3.21.0      # via onepush
pydantic==2.10.0          # via -r requirements-in.txt, fastapi
pydantic-core==2.27.0     # via pydantic
pyelftools==0.30          # via apkutils2
pygments==2.17.2          # via rich
pyparsing==3.1.1          # via packaging
pyreadline3==3.4.1        # via humanfriendly
pywin32==306              # via -r requirements-in.txt
pyyaml==6.0               # via -r requirements-in.txt
pyzmq==25.1.2             # via zerorpc
requests==2.31.0          # via adbutils, onepush, ppocr-onnx, uiautomator2
retry==0.9.2              # via adbutils, uiautomator2
rich==13.3.5              # via -r requirements-in.txt
shapely==2.0.2            # via ppocr-onnx
six==1.16.0               # via adbutils, anytree, uiautomator2
sniffio==1.3.0            # via anyio
starlette==0.27.0         # via fastapi
sympy==1.12               # via onnxruntime
tqdm==4.65.0              # via -r requirements-in.txt
typing-extensions==4.12.2  # via fastapi, pydantic, pydantic-core, uvicorn
uiautomator2==2.16.17     # via -r requirements-in.txt
uiautomator2cache==0.3.0.1  # via -r requirements-in.txt
urllib3==2.1.0            # via requests
uvicorn==0.23.2           # via -r requirements-in.txt
websockets==11.0.3        # via -r requirements-in.txt
whichcraft==0.6.1         # via adbutils, uiautomator2
wrapt==1.15.0             # via -r requirements-in.txt, deprecated
xmltodict==0.13.0         # via apkutils2
zerorpc==0.6.3            # via -r requirements-in.txt
zope-event==5.0           # via gevent
zope-interface==6.1       # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
